Dev: Multiple changes...:
- Add UI frontend for assets cleaning (with cleaned info dialog)

- Move update_launcher_settings to another module
