+++ quantum_launcher/src/state/message.rs
@@ -285,10 +285,15 @@ pub enum LauncherSettingsMessage {
     UiScaleApply,
     UiOpacity(f32),
     UiIdleFps(f64),
+
     ClearJavaInstalls,
     ClearJavaInstallsConfirm,
     ClearDownloadCache,
     ClearDownloadCacheConfirm,
+
+    CleanAssets, // Non-destructive
+    CleanAssetsFinished(Res<u64>),
+
     DefaultMinecraftWidthChanged(String),
     DefaultMinecraftHeightChanged(String),
     Rpc(RpcMessage),
